+++ src/dataproc/index.ts
@@ -1,0 +1,9 @@
+import { requireClusterEncryptionConfig } from './requireClusterEncryptionConfig'
+import { requireClusterInternalIpOnly } from './requireClusterInternalIpOnly'
+import { requireMetastoreServiceEncryptionConfig } from './requireMetastoreServiceEncryptionConfig'
+
+export const dataprocPolicies = [
+	requireClusterEncryptionConfig,
+	requireClusterInternalIpOnly,
+	requireMetastoreServiceEncryptionConfig
+]